requirements.txt
flask==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
openai>=1.0.0
tiktoken==0.3.0
aiohttp==3.8.1
selenium==4.9.0
webdriver-manager==3.8.6
shlex==0.0.3 